[NEW] freestore_free:  Replace new by linking only
The benefit of exorcising exception runtime bloat can be had without the
need to modify source (which otherwise must be in C++); just `use` this.

+++ base/iota/iota/freestore_free.hh
@@ -30,7 +30,7 @@
 
 extern void* operator_new_is_undefined;
 
-void* operator new( size_t )
+void* operator new( unsigned long )
 {
 	return operator_new_is_undefined;
 }
